test: extract EnforceTrackedSkillsInternalRule to its own file with docs
