+++ packages/camera/camera_android_camerax/test/validate_skills_test.dart
@@ -1,55 +1,11 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:dart_skills_lint/dart_skills_lint.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:logging/logging.dart';
 
-class EnforceTrackedSkillsInternalRule extends SkillRule {
-  @override
-  String get name => 'enforce-tracked-skills-internal';
-
-  @override
-  AnalysisSeverity get severity => AnalysisSeverity.error;
-
-  @override
-  Future<List<ValidationError>> validate(SkillContext context) async {
-    // Check if any files in the skill directory are tracked in git
-    final ProcessResult processResult = await Process.run('git', ['ls-files', context.directory.path]);
-    final String output = (processResult.stdout as String).trim();
-    if (output.isEmpty) {
-      // Not tracked by git, no enforcement needed
-      return [];
-    }
-
-    final Object? yaml = context.parsedYaml;
-    if (yaml == null) {
-      return [
-        ValidationError(
-          ruleId: name,
-          severity: severity,
-          file: 'SKILL.md',
-          message: 'Tracked skills must have "metadata: internal: true" in SKILL.md frontmatter.',
-        )
-      ];
-    }
-
-    final Object? metadata = (yaml as Map)['metadata'];
-    if (metadata is! Map || metadata['internal'] != true) {
-      return [
-        ValidationError(
-          ruleId: name,
-          severity: severity,
-          file: 'SKILL.md',
-          message: 'Tracked skills must have "metadata: internal: true" in SKILL.md frontmatter.',
-        )
-      ];
-    }
-
-    return [];
-  }
-}
+import 'enforce_tracked_skills_internal_rule.dart';
 
 void main() {
   test('Validate Skills', () async {
